- change structure - replace print with logger - add config file to run local and docker-compose - update readme
