+++ .gitignore
@@ -1,3 +1,4 @@
 temp/
 /src.tar.gz
 db_data/
+/local_files/
