+++ requirements.txt
@@ -1,0 +1,7 @@
+Flask==1.1.2
+pandas==1.0.0
+numpy==1.18.1
+fuzzywuzzy==0.17.0
+sqlalchemy
+psycopg2-binary==2.8.4
+python-Levenshtein-wheels
