--- src/Dockerfile
@@ -1,33 +1,0 @@
-FROM tiangolo/uwsgi-nginx-flask:python3.7
-
-ENV FLASK_APP api_server/api_app.py
-ENV FLASK_RUN_HOST 0.0.0.0
-
-# Indicate where uwsgi.ini lives
-ENV UWSGI_INI uwsgi.ini
-
-RUN apt-get update
-
-COPY requirements.txt /
-RUN pip install --no-cache-dir -r /requirements.txt
-
-# Set the folder where uwsgi looks for the app
-WORKDIR /paws-data-pipeline/
-
-# Copy the app contents to the image
-COPY api_server/static/ /app/static/
-RUN mkdir /app/static/uploads \
-          /app/static/uploads/current \
-          /app/static/output \
-          /app/static/output/reports \
-          /app/static/logs
-RUN chmod -R 777 /app/static
-
-COPY . /paws-data-pipeline/
-RUN chmod 777 /paws-data-pipeline
-
-# If you have additional requirements beyond Flask (which is included in the
-# base image), generate a requirements.txt file with pip freeze and uncomment
-# the next three lines.
-
-CMD ["flask", "run"]
