--- src/api_server/static/css/app.css
@@ -1,2 +1,0 @@
-.info {color: green}
-.error {color: red}
