--- src/docker-compose.yml
@@ -1,29 +1,0 @@
-version: '3'
-services:
-  db:
-    container_name: paws-compose-db
-    image: postgres:11.3-alpine
-    volumes:
-      - postgres:/var/lib/postgresql/data
-    environment:
-      POSTGRES_PASSWORD: thispasswordisverysecure
-    ports:
-      - "5432:5432"
-  web:
-    build: .
-    # Unfortunately docker-compose doesn't let us do much
-    # with regard to retroactively running commands, so we have
-    # to run a startup script, which loads paws db and then
-    # starts our service
-    #command: bash /paws-data-pipeline/ops_data/runserver.sh
-    ports:
-      - "5000:5000"
-    depends_on:
-      - db
-    environment: 
-      FLASK_ENV: development
-    container_name: paws-compose-web
-#using named volumes fixs a windows docker bug relating to container permissions
-#https://stackoverflow.com/questions/49148754/docker-container-shuts-down-giving-data-directory-has-wrong-ownership-error-wh
-volumes:
-  postgres:
