--- src/requirements.txt
@@ -1,6 +1,0 @@
-pandas==1.0.0
-numpy==1.18.1
-fuzzywuzzy==0.17.0
-python-Levenshtein
-sqlalchemy
-psycopg2-binary==2.8.4
