+++ src/scripts/flow_script.py
@@ -1,13 +1,8 @@
-import sys
 import os
 
-# get scripts folder to relative path
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 from scripts import load_paws_data, match_data, create_master_df
-
-CURRENT_SOURCE_FILES_PATH = '/app/static/uploads/current'
-UPLOADED_FILES_PATH = '/app/static/uploads/'
-OUTPUT_PATH = "/app/static/output/"
+from config import CURRENT_SOURCE_FILES_PATH, LOGS_PATH
+from flask import current_app
 
 MAPPING_FIELDS = {
     'salesforcecontacts': {
@@ -37,14 +32,12 @@ def start_flow():
         for uploaded_file in os.listdir(CURRENT_SOURCE_FILES_PATH):
             file_path = os.path.join(CURRENT_SOURCE_FILES_PATH, uploaded_file)
             file_name_striped = file_path.split('/')[-1].split('-')[0]
-            print('running load_paws_data on: ' + uploaded_file)
+            current_app.logger.info('running load_paws_data on: ' + uploaded_file)
             db_engine = load_paws_data.load_to_postgres(file_path, file_name_striped, True)
             pandas_tables[file_name_striped] = match_data.read_from_postgres(db_engine, file_name_striped)
             pandas_tables[file_name_striped] = match_data.cleanup_and_log_table(pandas_tables[file_name_striped],
                 MAPPING_FIELDS[file_name_striped],
                                                                                 'excluded_' + file_name_striped + '.csv')
-
-
 
         with db_engine.connect() as connection:
             create_master_df.main(connection)
@@ -55,7 +48,7 @@ def start_flow():
                 .pipe(match_data.match_cleaned_table, pandas_tables['volgistics'], 'volgistics', 'unmatched_volgistics.csv')
         )
 
-        matched_df.to_csv(os.path.join(match_data.LOG_PATH, 'matches.csv'), index=False)
+        matched_df.to_csv(os.path.join(LOGS_PATH, 'matches.csv'), index=False)
 
         # db_engine.dispose()  # we could close the db engine here once we're done with everything, but then it will be completely closed
         # See https://docs.sqlalchemy.org/en/13/core/connections.html#engine-disposal for design considerations.
