+++ src/static/css/app.css
@@ -1,0 +1,2 @@
+.info {color: green}
+.error {color: red}
