--- src/uwsgi.ini
@@ -1,7 +1,0 @@
-[uwsgi]
-module = api_server.api_app
-callable = app
-uid = 1000
-master = true
-threads = 2
-processes = 4
